+++ lib/src/data/api_response.dart
@@ -7,6 +7,21 @@ class ApiResponse extends http.Response {
   /// Time when [ApiResponse] was created.
   final DateTime received;
 
+  /// Represent reponse success
+  ///
+  /// [statusCode] is in the range from `200` to `299`, inclusive.
+  bool get ok => statusCode >= 200 && statusCode <= 299;
+
+  /// Represent redirect status.
+  ///
+  /// When [statusCode] equals `301`, `302`, `303`, `307`, or `308`.
+  bool get redirect =>
+      statusCode == 301 ||
+      statusCode == 302 ||
+      statusCode == 303 ||
+      statusCode == 307 ||
+      statusCode == 308;
+
   ApiResponse.fromHttp(http.BaseRequest request, http.Response response,
       Map<String, dynamic> linkData)
       : this.linkData = linkData ?? <String, dynamic>{},