+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [0.5.0-dev.1] - 24 July 2020.
 * Added experimental cache support.
+## [0.4.2] - 03 June 2020.
+* Added `ok` property to `ApiResponse`
+* Added `redirect` property to `ApiResponse`
 ## [0.4.1] - 27 April 2020.
 * Added `FileField.fromStream` constructor.
 ## [0.4.0] - 19 April 2020.